+++ src/IntegrationSchema.ts
@@ -299,6 +299,10 @@ import DatabaseJson from './schemas/Database.json';
 export const Database = DatabaseJson;
 IntegrationSchema.addSchema(Database);
 
+import DataTableJson from './schemas/DataTable.json';
+export const DataTable = DataTableJson;
+IntegrationSchema.addSchema(DataTable);
+
 import DataStoreJson from './schemas/DataStore.json';
 export const DataStore = DataStoreJson;
 IntegrationSchema.addSchema(DataStore);