+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- [#96](https://github.com/JupiterOne/data-model/issues/96) - Introduce
+  `DataTable` entity class
+
 ## 0.23.0 - 2021-05-24
 
 ### Added